requirements.txt
Django==2.1.9
python-decouple==3.1
django-debug-toolbar==2.0
pipenv==2018.11.26
